Revert "bug: bad solution but somehow works?"
This reverts commit 6d6b4ade11b3b926f457452114044282be3c096b.
